Support for SConscript subdir for cscope builder.
Add new python list subclass, backed by a generator for providing the
list data.

Add functions to translate executable name so that command invocations
(impling a search on PATH) still load the same executable after changing
current directory.

Update cscope builder to use CSCOPEDIRECTORY variable for the directory
to run `cscope` command in. Inputs are translated to be relative to this
directory, default is the  current source directory.